+++ test/modern_bpf/test_suites/syscall_exit_suite/chmod_x.cpp
@@ -12,7 +12,7 @@ TEST(SyscallExit, chmodX)
 	const char* filename = "*//null";
 	uint32_t mode = S_IXUSR;
 	assert_syscall_state(SYSCALL_FAILURE, "chmod", syscall(__NR_chmod, filename, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -31,7 +31,7 @@ TEST(SyscallExit, chmodX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO) */
+	/* Parameter 1: res (type: PT_ERRNO32) */
 	evt_test->assert_numeric_param(1, errno_value);
 
 	/* Parameter 2: filename (type: PT_FSPATH) */
