+++ test/modern_bpf/test_suites/syscall_exit_suite/fchmod_x.cpp
@@ -12,7 +12,7 @@ TEST(SyscallExit, fchmodX)
 	int32_t mock_fd = -1;
 	uint32_t mode = S_IXUSR;
 	assert_syscall_state(SYSCALL_FAILURE, "fchmod", syscall(__NR_fchmod, mock_fd, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -31,11 +31,11 @@ TEST(SyscallExit, fchmodX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO) */
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: res (type: PT_ERRNO32) */
+	evt_test->assert_numeric_param(1, errno_value);
 
-	/* Parameter 2: fd (type: PT_FD) */
-	evt_test->assert_numeric_param(2, (int64_t)mock_fd);
+	/* Parameter 2: fd (type: PT_FD32) */
+	evt_test->assert_numeric_param(2, mock_fd);
 
 	/* Parameter 3: mode (type: PT_MODE) */
 	evt_test->assert_numeric_param(3, (uint32_t)PPM_S_IXUSR);
