+++ test/modern_bpf/test_suites/syscall_exit_suite/fchmodat_x.cpp
@@ -14,7 +14,7 @@ TEST(SyscallExit, fchmodatX)
 	uint32_t mode = S_IXUSR;
 	uint32_t flags = 0;
 	assert_syscall_state(SYSCALL_FAILURE, "fchmodat", syscall(__NR_fchmodat, mock_dirfd, pathname, mode, flags));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -33,11 +33,11 @@ TEST(SyscallExit, fchmodatX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO) */
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: res (type: PT_ERRNO32) */
+	evt_test->assert_numeric_param(1, errno_value);
 
-	/* Parameter 2: dirfd (type: PT_FD) */
-	evt_test->assert_numeric_param(2, (int64_t)mock_dirfd);
+	/* Parameter 2: dirfd (type: PT_FD32) */
+	evt_test->assert_numeric_param(2, mock_dirfd);
 
 	/* Parameter 3: filename (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(3, pathname);
