+++ test/modern_bpf/test_suites/syscall_exit_suite/mkdir_x.cpp
@@ -12,7 +12,7 @@ TEST(SyscallExit, mkdirX)
 	uint32_t mode = 7;
 	const char* path = "*//null";
 	assert_syscall_state(SYSCALL_FAILURE, "mkdir", syscall(__NR_mkdir, path, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -31,7 +31,7 @@ TEST(SyscallExit, mkdirX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO)*/
+	/* Parameter 1: res (type: PT_ERRNO32)*/
 	evt_test->assert_numeric_param(1, errno_value);
 
 	/* Parameter 2: path (type: PT_FSPATH) */
