+++ test/modern_bpf/test_suites/syscall_exit_suite/mkdirat_x.cpp
@@ -13,7 +13,7 @@ TEST(SyscallExit, mkdiratX)
 	const char* path = "/invalid/path";
 	uint32_t mode = 8;
 	assert_syscall_state(SYSCALL_FAILURE, "mkdirat", syscall(__NR_mkdirat, mock_dirfd, path, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -32,11 +32,11 @@ TEST(SyscallExit, mkdiratX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO) */
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: res (type: PT_ERRNO32) */
+	evt_test->assert_numeric_param(1, errno_value);
 
-	/* Parameter 2: dirfd (type: PT_FD) */
-	evt_test->assert_numeric_param(2, (int64_t)mock_dirfd);
+	/* Parameter 2: dirfd (type: PT_FD32) */
+	evt_test->assert_numeric_param(2, mock_dirfd);
 
 	/* Parameter 3: path (type: PT_FSRELPATH) */
 	evt_test->assert_charbuf_param(3, path);
