+++ test/modern_bpf/test_suites/syscall_exit_suite/open_x.cpp
@@ -45,8 +45,8 @@ TEST(SyscallExit, openX_success)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: ret (type: PT_FD)*/
-	evt_test->assert_numeric_param(1, (int64_t)fd);
+	/* Parameter 1: ret (type: PT_FD32)*/
+	evt_test->assert_numeric_param(1, fd);
 
 	/* Parameter 2: name (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(2, pathname);
@@ -85,7 +85,7 @@ TEST(SyscallExit, openX_failure)
 	int flags = O_RDWR | O_TMPFILE | O_DIRECTORY;
 	mode_t mode = 0;
 	assert_syscall_state(SYSCALL_FAILURE, "open", syscall(__NR_open, pathname, flags, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL  ===========================*/
 
@@ -104,8 +104,8 @@ TEST(SyscallExit, openX_failure)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: ret (type: PT_FD)*/
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: ret (type: PT_FD32)*/
+	evt_test->assert_numeric_param(1, errno_value);
 
 	/* Parameter 2: name (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(2, pathname);
