+++ test/modern_bpf/test_suites/syscall_exit_suite/openat_x.cpp
@@ -46,11 +46,11 @@ TEST(SyscallExit, openatX_success)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: fd (type: PT_FD) */
-	evt_test->assert_numeric_param(1, (int64_t)fd);
+	/* Parameter 1: fd (type: PT_FD32) */
+	evt_test->assert_numeric_param(1, fd);
 
-	/* Parameter 2: dirfd (type: PT_FD) */
-	evt_test->assert_numeric_param(2, (int64_t)PPM_AT_FDCWD);
+	/* Parameter 2: dirfd (type: PT_FD32) */
+	evt_test->assert_numeric_param(2, PPM_AT_FDCWD);
 
 	/* Parameter 3: name (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(3, pathname);
@@ -90,7 +90,7 @@ TEST(SyscallExit, openatX_failure)
 	int flags = O_RDWR | O_TMPFILE | O_DIRECTORY;
 	mode_t mode = 0;
 	assert_syscall_state(SYSCALL_FAILURE, "openat", syscall(__NR_openat, dirfd, pathname, flags, mode));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL  ===========================*/
 
@@ -109,11 +109,11 @@ TEST(SyscallExit, openatX_failure)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: fd (type: PT_FD) */
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: fd (type: PT_FD32) */
+	evt_test->assert_numeric_param(1, errno_value);
 
-	/* Parameter 2: dirfd (type: PT_FD) */
-	evt_test->assert_numeric_param(2, (int64_t)PPM_AT_FDCWD);
+	/* Parameter 2: dirfd (type: PT_FD32) */
+	evt_test->assert_numeric_param(2, PPM_AT_FDCWD);
 
 	/* Parameter 3: name (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(3, pathname);
