+++ test/modern_bpf/test_suites/syscall_exit_suite/rmdir_x.cpp
@@ -11,7 +11,7 @@ TEST(SyscallExit, rmdirX)
 
 	const char* path = "*//null";
 	assert_syscall_state(SYSCALL_FAILURE, "rmdir", syscall(__NR_rmdir, path));
-	int64_t errno_value = -errno;
+	int32_t errno_value = -errno;
 
 	/*=============================== TRIGGER SYSCALL ===========================*/
 
@@ -30,8 +30,8 @@ TEST(SyscallExit, rmdirX)
 
 	/*=============================== ASSERT PARAMETERS  ===========================*/
 
-	/* Parameter 1: res (type: PT_ERRNO)*/
-	evt_test->assert_numeric_param(1, (int64_t)errno_value);
+	/* Parameter 1: res (type: PT_ERRNO32)*/
+	evt_test->assert_numeric_param(1, errno_value);
 
 	/* Parameter 2: path (type: PT_FSPATH) */
 	evt_test->assert_charbuf_param(2, path);
